backend/requirements.txt
djangorestframework==3.1.8
pillow==8.2.0
crispy-forms-gds==0.2.2
html5lib==1.1
wheel==0.36.2
JPype1==1.3.0
konlpy==0.5.2
wordcloud==1.8.1